+++ jar-filter/src/test/resources/delete-constructor/kotlin/net/corda/gradle/PrimaryConstructorsToDelete.kt
@@ -1,5 +1,4 @@
-@file:JvmName("PrimaryConstructorsToDelete")
-@file:Suppress("UNUSED")
+@file:Suppress("unused", "PackageDirectoryMismatch")
 package net.corda.gradle
 
 import net.corda.gradle.jarfilter.DeleteMe