+++ jar-filter/src/test/resources/delete-overloaded-function/kotlin/net/corda/gradle/HasOverloadedFunction.kt
@@ -1,4 +1,4 @@
-@file:Suppress("unused")
+@file:Suppress("unused", "PackageDirectoryMismatch")
 package net.corda.gradle
 
 import net.corda.gradle.jarfilter.DeleteMe