+++ jar-filter/src/test/resources/delete-val-property/kotlin/net/corda/gradle/HasValPropertyForDelete.kt
@@ -1,4 +1,4 @@
-@file:Suppress("UNUSED")
+@file:Suppress("unused", "PackageDirectoryMismatch")
 package net.corda.gradle
 
 import net.corda.gradle.jarfilter.DeleteMe