+++ jar-filter/src/test/resources/stub-function/kotlin/net/corda/gradle/RuntimeAnnotations.kt
@@ -1,10 +1,8 @@
-@file:JvmName("RuntimeAnnotations")
+@file:Suppress("PackageDirectoryMismatch")
 package net.corda.gradle
 
 import kotlin.annotation.AnnotationRetention.*
 import kotlin.annotation.AnnotationTarget.*
-import kotlin.annotation.Retention
-import kotlin.annotation.Target
 
 @Target(VALUE_PARAMETER)
 @Retention(RUNTIME)