+++ jar-filter/src/test/resources/stub-var-property/kotlin/net/corda/gradle/HasVarPropertyForStub.kt
@@ -1,5 +1,4 @@
-@file:JvmName("HasVarPropertyForStub")
-@file:Suppress("UNUSED")
+@file:Suppress("unused", "PackageDirectoryMismatch")
 package net.corda.gradle
 
 import net.corda.gradle.jarfilter.StubMeOut